Buffer pool with comments addressed. Summary 1. Modified `ByteBufferPool` interface (dropped two methods) 2. In both `SimpleByteBufferPool` and `FifoByteBufferPool`, Simplify `deallocate()` by just ensuring pool's available memory does not go beyond its capacity 3. Modified unit tests to eliminate `Thread.sleep()` 4. Detailed documentation 5. Test coverage is 100%.
Test done: ./gradlew build

Primary reviewer: Priyesh, Gopal

Note: There was a discussion how to design the 'ByteBufferPool' interface, and the current agreement is to keep it as simple as possible by keeping  only `allocate()` and `deallocate()`. Other methods such as `availableMemory()` and `capacity()` that might be useful for statistics collection and system fine tune can be left in implementations.

The reason for this is that, a buffer pool may be unbounded, so that `capacity()` and `availableMemory()` will be meaningless for these pools. It makes little sense to keep these two methods in the interface. However, a full solution will be that there can be a `BoundedByteBufferPool` interface, which has the two methods, and the current `SimpleByteBufferPool` and `FifoByteBufferPool` should actually implement the `BoundedByteBufferPool` interface. To avoid over-complicating the current "simple" work, we decided to keep this for future enhancement.