Operation policy for non-blocking router
This PR introduces OperationPolicy for non-blocking router. After
several rounds of discussion, it is preferred to abstract the key
methods from general policy instances, and wrap them as an interface.
For simplicity, an policy instance maintains the status of an operation.

The new OperationPolicy is not meant to replace the old operation
policy, and hence they cannot be used interchangeably.

The PR includes:
1. An OperationPolicy interface that defines key methods;
2. An AmbryOperationPolicy that implements OperationPolicy;
3. A PolicyConfig that specifies key parameters with default values;
4. Unit test

Test done: ./gradlew build

Primary reviewer: Priyesh & Siva