Keeping connections alive on error responses.
The current code closes connections on all error responses which is wasteful.
With these changes, we keep connections alive when we have errors on non-POST requests.

Also fixes a bug where a HEAD request for a deleted blob returned a body (HEAD is not supposed to return a body).
Also adds some more keep-alive tests.