Introduces metrics framework for NB Router.
This patch introduces a metrics framework for NB Router, and it is rebased on PR/327. It has two major parts:
1. General metrics and metrics-updating methods at the router level. It makes clear and avoids missing general metrics by using the general metrics-updating method (changed the way suggested from a previous discussion to update metrics by directly referencing the individual metrics);
2. General metrics and metrics-updating methods at the NodeId level. This will help to conduct root-cause analysis by tracking metrics of an individual request to a specific node.
3. Collect metrics for Put and Delete.

This is intended to be an early effort for router metrics, and I tried to limit the size of this PR. More metrics are yet to be collected, and new PR will be made to update each.

Reviewer: Priyesh & Siva
Estimated review time: 40min
Test: No specific tests are made for this patch, however ./gradlew build passes.