Adding getStatus() and getHeader() APIs in RestResponseChannel (#319)
The RestResponseChannel is passed around as a piece of data and can help avoid
repeated code by simply exposing the headers that have been set. Components
that pass a RestResponseChannel around can then start communicating by the
presence, absence and value of headers.

This change also adds and refines some tests in NettyResponseChannel.