Add support for getBlob() in the router (#308)
A GetBlobOperation makes use of GetChunks responsible for fetching
chunks of a blob. These GetChunks are reusable (similar to PutChunks)
and can cycle through multiple chunks of a blob.

The first chunk of a blob is treated as a special chunk as that
needs to be fetched and interpreted to determine whether the blob is
simple or composite. After the first, if the blob is composite,
subsequent chunks are fetched in parallel (limited by MAX_IN_MEM_CHUNKS).

As soon as the first bytes of the data arrives, the operation callback
is invoked so that the data can be streamed out. Any exception
encountered is notified to the caller via the read callbacks.