Changing invalid request state detection logic in NettyMessageProcessor.
Previously NettyMessageProcessor used to be ready for subsequent requests on the same channel
only after NettyResponseChannel reported that responses had been fully sent out (including
sending of LastHttpContent  which is an artifact of Netty and has no meaning in responses
with Content-Length set). This created a race condition between the server and client in the
scenario where clients detect that they have fully received a response and send another request
while the server still hadn't realized that the response has been fully sent out. This race
condition became more pronounced and started inducing errors when the NonBlockingRouter was used.

This commit fixes this problem by declaring the NettyMessageProcessor ready for subsequent requests
as soon as the current request is fully received from the client. This way, there are no race conditions
because NettyMessageProcessor is ready even before the response sending has started. The only way this
can cause an error is if the client sends a request without waiting for the previous response in which
case that would be a bug with the client and not the server.