Fixes to LastHttpContent generation in NettyResponseChannel
Before this change, LastHttpContent was always empty and written to channel once all
the real data had been written. With this change, empty LastHttpContent is written only
in the case of responses with "Transfer-Encoding : Chunked". If a response has Content-Length
set, we now send back either a FullHttpResponse (in case Content-Length is 0) or make the last
chunk LastHttpContent (if Content-Length > 0).

This continues the fix introduced in #345 and fixes another race condition. This also solves #346.
The race condition described in #346 can be detrimental if a new request arrives and the response to
that request starts being sent before the LastHttpContent of the previous request has been written out.