Performance optimizations for the router (#386)
This patch introduces two main changes: a) Adding a deterministic
wait-notify flow for the ChunkFiller thread to go to sleep and get
notified whenever there is work to be done. b) Adding a "wakeUp" for the
NetworkClient that will in turn wake up the associated Selector from its
poll if an operation manager becomes poll-eligible asynchronously to the
RequestResponseHandler thread.

Also adds a couple of other fixes.