Make the router resilient to unexpected errors
Ensure that exceptions are not propagated outside of the operation
manager poll(), handleResponse() and the networkClient sendAndPoll()
methods, as these can bring down the router. Exceptions are unexpected
in these methods, so track them via metrics but do not aggressively
shut down the router, as there is a high chance that these are
recoverable errors, or errors isolated to a particular operation.