Ensure operation tracker does not eliminate down replicas (#424)
Operation tracker should not avoid returning down replicas. That is
a decision to be made by the router, if at all. Operation tracker
should however ensure that down replicas are ordered last.

This will ensure that even if the failure detector marks all partitions
as unhealthy (which can happen if servers are all restarted within a
short timespan for example), operations have a chance to succeed.