Adding backpressure support to NettyRequest
This change adds backpressure support to NettyRequest allowing it to suspend (and resume) reading
from the client based on a configurable watermark. This helps keep memory usage at the frontend in
check.

This change also adds a test for the corner case where the underlying buffer from Netty's ByteBuf
cannot be extracted and the data has to be copied.