Write out put requests efficiently
Put Requests are written out inefficiently today - by creating a buffer
that is only large enough to hold the request header, blob properties and
user metadata and then reusing the same buffer to read in the blob data
part by part from the InputStream and writing it out iteratively.
This is particularly inefficient now, as the router uses a selector and all
PutRequests get written out iteratively. This is also unnecessary now as
the data is already available in a ByteBuffer, so it can be written out
efficiently without any extra memory allocations.

Considerable latency improvements were seen in perf tests with this change.