Refactor, address review comments.
Make a cleaner separation between a PutRequest that is sent out and one
that is received. A received request should not be Sendable (does not
implement Send). This makes things easier internally as well, so that an
InputStream is maintained only for received requests (as that comes in
via a SocketInputStream or whatever) - whereas when PutRequests are sent
out, we get to assume that there is always an accompanying materialized
blob buffer.

Also make a few changes to remove those coordinator tests that would
have required changes due to this patch.