Write out put requests efficiently (#445)
Put Requests are written out inefficiently today - by creating a buffer
that is only large enough to hold the request header, blob properties and
user metadata and then reusing the same buffer to read in the blob data
part by part from the InputStream and writing it out iteratively.
This is particularly inefficient now, as the router uses a selector and all
PutRequests get written out iteratively. This is also unnecessary now as
the data is already available in a ByteBuffer, so it can be written out
efficiently without any extra memory allocations.

Considerable latency improvements were seen in perf tests with this change.

Also, making a cleaner separation between a PutRequest that is sent out and one
that is received. A received request should not be Sendable (does not
implement Send). This makes things easier internally as well, so that an
InputStream is maintained only for received requests (as that comes in
via a SocketInputStream or whatever) - whereas when PutRequests are sent
out, we get to assume that there is always an accompanying materialized
blob buffer.