Changes to the Log to support log segmentation
This commit introduces the changes required for the log to use LogSegment instances. It allows
the log to be segmented into multiple segments.
However, the index, hard delete and BlobStore do not yet completely support segmentation and hence
the log cannot yet be segmented.
The current configs make sure that segments aren't created