Using store abstractions in tools + supporting segmented logs in tools (#509)
ambry-tools had some repeated code for reading index segments. This commit makes
use of the IndexSegment class to read the index file and thus unifies the code that
enables interpretation of index segment formats

This commit also enables support for segmented logs in ambry tools and enables them
to work with segmented or non segmented logs

This commit also removed the option to verify consistency of log entries with index.
This functionality is (almost) fully covered in the code that verifies the consistency
of index entries with the log and thus is no longer required.

This commit also removes the IndexConsistencyChecker tool which is no longer relevant and
has been replaced by better tools (compareIndexEntriesWithLog).