Adding CompactionDetails and CompactionLog
CompactionDetails - serves as a job spec that will define the segments under compaction. Medium of exchange
of info b/w the component that determines what needs to be done and the component that does the work.
CompactionLog - a recovery log that checkpoints the different stages of compaction and helps in recovery in
cases of restarts and crashes