Avoid upfront size limitation for router puts
Relaxing the requirement to provide the size in advance allows for truly
streaming puts, where a caller may not know the size of the stream in advance.
- The router depended heavily on the size, so this change is a little extensive.
Unlike before, when the router could determine that a stream has ended the moment
the last byte arrives, now it will know that only when the channel gets closed.
- The MetadataChunk object is always created now and will track all data chunks,
even for non-composite blobs. When it gets finalized, it will result in an actual
metadata blob only if there are more than one data chunks.
- For puts, the router will ignore the size in the passed in BlobProperties. The
BlobProperties that is sent to the server during a PUT and returned during GETs will
always have the correct size.