Making the index friendly to a switch of active index segments
Compaction needs to atomically change the valid index segments after the completion
of a cycle. PersistentIndex was built with the assumption that the reference to the
valid index segments will not change. This commit removes that assumption and makes
all functions in the index work with a view of the valid index segments at a certain
point in time