Wake up RequestResponseHandler when the last chunk becomes ready
For puts, the RequestResponseHandler should be woken up whenever chunks become
ready to be sent out to the server. The wake-up is done by calling onPollReady().
This is missing in a newly added code path (PR #540) for when the last chunk becomes
ready. This considerably increased the Put latency. This patch fixes that.