Introduces Notifier TopicListener
This PR introduces Notifier and TopicListener as a means for
communications among different end points. A TopicListener can subscribe
to a topic through a Notifier. A message of a topic can be sent
through a Notifier, such that all the TopicListeners subscribed to that
topic can receive the message. The PR comes with a Helix-based Notifier
implementation.

This PR is split from PR #624 to ease the reviewing work. PR #624 is
still kept open as a reference to the final goal, and will be rebased on
each smaller PRs and merged the last.

Test done: ./gradlew clean && ./gradlew build
Unit tests are made against MockHelixPropertyStore generated through
MockHelixPropertyStoreFactory. However, tests also passed against a
HelixPropertyStore backed by a locally deployed ZooKeeper. Due to the
long time to instantiate a real ZooKeeper server in code, this part will
be made in an integration test in the future.

Test coverage:
Class Class, %  Method, % Line, %
HelixNotifier 100% (2/ 2) 92.3% (12/ 13)  96.8% (60/ 62)
HelixPropertyStoreFactory 100% (1/ 1) 50% (1/ 2)  18.2% (2/ 11)

The low test coverage for HelixPropertyStoreFactory is due to using
MockHelixPropertyStoreFactory.