Assigns id to each datacenter.
This patch brings in an internal concept of datacenter id, and generates
an id in short for each datacenter. Querying the id of a datacenter is
supported by the ClusterMap interface.

The datacenter id will be needed when it is embeded into blob id, to
indicate the originating datacenter of the blob.

This patch also introduces a util method to check if a datacenter name
is valid to add to a cluster.

Test done: ./gradlew clean && ./gradlew build

Test coverage:

Class Class, %  Method, % Line, %
ClusterMapMetrics 100% (21/ 21) 39% (23/ 59)  61.7% (79/ 128)
ClusterMapUtils 100% (1/ 1) 91.7% (11/ 12)  91.7% (44/ 48)
CompositeClusterManager 100% (1/ 1) 100% (14/ 14) 88.8% (87/ 98)
HardwareLayout  100% (1/ 1) 88.5% (23/ 26)  89.7% (130/ 145)
HelixBootstrapUpgradeTool 0% (0/ 1) 0% (0/ 2) 0% (0/ 45)
HelixBootstrapUpgradeUtil 100% (1/ 1) 84.2% (16/ 19)  76.6% (232/ 303)
HelixClusterManager 100% (5/ 5) 95.2% (40/ 42)  96.2% (254/ 264)
HelixClusterManagerMetrics  100% (16/ 16) 92.1% (35/ 38)  97.2% (103/
106)
StaticClusterManager  100% (2/ 2) 91.4% (32/ 35)  80.5% (178/ 221)