BlobIdV2
This patch introduces BlobId v2 that embeds accountId, containerId, and
datacenterId. The first two will be an in-blobId identifier for the
ownership of the blob, and the third will be used to get the datacenter
where the blob was originally posted.

To include the three additional pieces of information, the blobId
version needs to be increased from 1 to 2. To better support multiple
versions, a builder class is introduced to build blobIds in different
versions.

Test done:
./gradlew clean && ./gradlew build

Test coverage:
Class Class, %  Method, % Line, %
BlobId  100% (1/ 1) 88.2% (15/ 17)  80.4% (78/ 97)
BlobIdBuilder 100% (1/ 1) 100% (7/ 7) 100% (16/ 16)