Inflates Account and Container from Json Metadata (#628)
This PR introduces versioning and understanding of account/container
metadata, and inflates Account and Container objects from Json metadata.

This PR is split from PR #624 to ease the reviewing work. PR #624 is
still kept open as a reference to the final goal, and will be rebased on
each smaller PRs and merged the last.