Introduces Notifier and TopicListener (#629)
This PR introduces Notifier and TopicListener as a means for
communications among different end points. A TopicListener can subscribe
to a topic through a Notifier. A message of a topic can be sent
through a Notifier, such that all the TopicListeners subscribed to that
topic can receive the message. The PR comes with a Helix-based Notifier
implementation.