Add support for dynamically updating partition sealed state. (#644)
Partition's sealed state is dependent on the sealed states of its
replicas (which is controlled individually by the respective nodes).
A partition is READ_WRITE if *all* its replicas are READ_WRITE. Equivalently,
a partition is READ_ONLY if *any* of its replicas is READ_ONLY.

Individual instances (or admin operations) are supposed to update the
InstanceConfig to mark their replicas as READ_ONLY or READ_WRITE (this is not
part of this patch). When this change is made, HelixClusterManager updates the
states of the replicas as part of the InstanceConfig change handling. Changes to
replica sealed state is tracked at the highest level and a getPartitionState()
call on a Partition will refresh and cache its state from those of its replicas
if there has been such a change since the last refresh.