Blob id v2 (#646)
This patch introduces BlobId v2 that embeds flags, accountId, containerId, and
datacenterId. The first two will be an in-blobId identifier for the
ownership of the blob, and the third will be used to get the datacenter
where the blob was originally posted.

To include the four additional pieces of information, the blobId
version needs to be increased from 1 to 2.

A major change is temporarily setting BlobId#CURRENT_VERSION to
BLOBID_V1, so that newly created Blob ids will continue to be in BLOBID_V1. BlobId#CURRENT_VERSION will be set to v2 when writing v2 is enabled.