Add AclService interfaces (#676)
This adds an AclService<P> interface. The service will make access control decisions based on a generic principal of type P, a Resource, and an Operation. It also supports creating/removing ACL rules for a resource. Resources that one wants to set access rules on must implement the AclService.Resource interface. This interface provides the ACL service with a unique way of identifying a resource.