CLI tool to create or update accounts.
This pr introduces a cli-based tool to operate on accounts. It takes a
file in json array, with each element to be an account in its json
format, and underlyingly employs AccountService to do the write
operation.

An integration test has been added to test against a locally-deployed
ZooKeeper server.

Test: ./gradlew build && ./gradlew test