Fixing a build failure due to runtime class dependencies
In #693, the InMemoryRouter was moved to the ambry-router module and
a test source dependency on ambry-router was added to ambry-rest.
However, a dependency on the main source was not added and InMemoryRouter
uses some code in the main source. This was causing a NoClassDefFoundError.
This commit adds a dependency for the tests on the main source of ambry-router
which should fix the problem