Adding a tool that can copy data from one store to another (#704)
The tool (StoreCopier) also allows transforming the messages in the store
during copy. The possible uses of this tool will be when messages in a
store need data format changes (for e.g. interconversion b/w segmented
and non-segmented logs or if more data needs to be added to the message).
Also adds a DiskReformatter tool that uses the StoreCopier to reformat the 
whole disk.