Improving DumpIndexTool - can be used as both standalone tool or library
Remodelling the functions in DumpIndexTool so that they can be used as APIs
to build more tool functionality on top of. Also supports most of the operations
that were being supported (correctly) previously. Support for getting "active"
blobs has been removed since the old support was incorrect and there is no use
case to add support for it right now. However, it should be easy to add if required
at a later date by using the verifyIndex() method.

Also made changes to the ConsistencyCheckerTool to isolate its functionality from
the DumpIndexTool.