Change MockHelixAccountServiceFactory to always use MockHelixPropertyStore (#751)
Previously the constructor took a fourth parameter
to allow caller to specify if should use mock helixpropertystore.  However,
this will be hard to plug in the MockHelixAccountServiceFactory to any
integration test that uses a mock property store. This PR introduces
this possibility. Usages that require an actual HelixPropertyStore can use
HelixAccountServiceFactory.

Test: clean build