Adds background thread to pull account updates from ZK (#742)
This PR adds a background thread that pulls the account update from ZK
in addition to being triggered by account update is made. This is to
avoid missing update due to unexpected behavior of listener, and add
confidence that after a period of configured time, it will always get in
sync with the remote accounts.