Setting a valid datacenter ID in the blob ID
This commit provides a way of assigning IDs to datacenters so that this
information is correctly embedded inside the blob ID. In effect, this
will provide information on which DC the blob was created in.

This information could be applied in many ways. Two immediate examples:
1. Cross colo GET requests - Know exactly which DC to proxy to to get the blob
which wasn't found in the local datacenter.
2. Track workload - how often is a blob fetched in a DC it was not created in?