Added option to disable reverse DNS lookups. (#795)
This commit makes the behavior of Ambry to use a reverse name lookup
configurable. By default Ambry (old and current behavior) will do a
reverse name lookup with DNS for every clustermap host. When the option
'clustermap.resolve.hostnames' is set to 'false' Ambry will trust the
hostname in the clustermap.