Adding a preprocess step in security service
This commit
1) Adds a function to pre-process a request in SecurityService
2) Invokes this function on all supported methods in AmbryBlobStorageService
3) Invokes processResponse() in GetSignedUrlHandler

This change is generally targetted to support the use case where some custom
processing of the request has to happen before any business logic is triggered.

Specifically the change is being executed now in order to support encryption of
signed URLs if required. The idea is that processResponse() will encrypt the query
parameters of a signed URL and preProcessRequest() will decrypt and insert the
parameters into the RestRequest when the signed url is presented for use.