Adding encryption support to Router (#801)
This patch adds encryption support to Router. Router relies on isEncrypted() in BlobProperties to determine if the blob needs to be encrypted or not during uploads. If set, both userMetadata and blob content will be encrypted. On Get path, the same will be decrypted before responding. In other words, the stream that comes out of router for a GetRequest will always match the plain content passed in.