Add support for BlobId crafting
Add support for generating crafted blob id for a given BlobId of
any version and type. The crafted id will have the same constituents
(including uuid) as the input id, except for the version (crafted ids
will be V3), type (crafted id will be of type CRAFTED), account and
container (which will be those that are passed in as parameters to
the craft() call).