Add support for parsing BlobId preamble without a clustermap
The preamble is internal to the BlobId class and is something
that can be parsed from a Blob id string without a cluster map.

This is used within newly introduced methods that determine
if a given blob id in string form is crafted, and get account
and container ids, both by parsing the preamble from the string.