Add support for data prefetch.
Add support for data prefetch (as opposed to transferTo which does
zero copy) for get operations that can be enabled via a config.
Data prefetching has the effect of disk I/O happening in the context
of the RequestHandler thread as opposed to the processor/selector
thread as in the case of transferTo.