Introducing a UpdateRecord
In order to support more general updates, this commit introduces the
concept of UpdateRecord. The first two versions of this record are
equivalent to DeleteRecord versions 1 and 2, The third version is
able to encapsulate different types of records.

This commit adds read support. A future commit will convert the
DeleteMessageFormatInputStream to write in the new format.