Resolves possible race conditions for PartitionsState on init
AmbryPartition is initialized with state = READ_WRITE. The state is
resolved on the first call to getPartitionState(). This may cause some
race conditions just after init causing the clustermap to return RO
partitions as RW if certain conditions are met. This change resolves
PartitionState before init is complete to avoid such race conditions