Introduce PutBlobOptions in router API
This will allow request options to be passed into the Router::putBlob
method. In contrast to the BlobProperties, which are persisted with
the blob in storage, PutBlobOptions only influence the router logic.

For now, the options object is empty, but will soon include some options
required for data chunk upload.