Handling a recovery case that wasn't handled before (#986)
The PUT record related to a TTL update record may be cleaned up
by compaction (due to a DELETE). In such a log, the previous code
would crash because it expected a TTL update to necessarily have
a PUT in the index. This commit fixes that case. Recovery now expects
that a TTL update without a preceding PUT necessarily has a succeeding
DELETE.