Supporting TTL updates in BlobStoreStats (#1011)
Supported TTL updates in BlobStoreStats for on demand scanning
The validity of a TTL update in the context of BlobStoreStats is
1. Valid if the PUT is still valid
2. Valid if the PUT is invalid but is in another log segment
3. Invalid otherwise (no PUT or PUT invalid and in the same log segment)

This is a little different from compaction which considers the whole range
of log segments being compacted to decide a case similar to #2. This will
necessitate a slight change to the StatsBasedCompactionPolicy to account
for this difference in behavior.

This commit does not address online updates (for bucketing). The feature
is not immediately required + has some pre-existing bugs (tests are already
disabled). This will be taken up as future work.