Prevent busy polling in replication thread (#1026)
This change make the thread sleep for a configurable amount of time before running the replication cycle again when replication is all caught up.
This is done to prevent busy polling which results in many unnecessary network calls. In addition, metrics are added for total bytes replicated per
execution and the percentage of partitions fetched that were actually meaningful.