[stitched uploads] Implement Router.stitchBlob() (#1081)
Add support for stitching blobs to the router. The main change is to
allow PutOperation to take in a list of chunk IDs (+ metadata) and
upload a freestanding metadata chunk. The caller of the stitchBlob
method is responsible for ensuring that the metadata in the ChunkInfo
object is correct.

For now, this will rely on the existing metadata format which requires
that all intermediate chunks are the same size. Future changes to this
format will allow for stitching chunks of arbitrary size.