[WIP] Populate correct blob size on getBlobInfo requests
When performing a getBlobInfo request, fetch the data segment of the
blob. If it is a metadata blob, use the size in the metadata content.
If it is an encrypted simple blob, use the size of the decrypted content
buffer.

Similarly, ensure that the blob size reflects the decrypted content size
on getBlob requests for a simple encrypted blob.