Make cluster manager initialization more resilient
Ensure that Helix cluster manager initializes and runs (albeit in a
degraded state) even if there are connectivity issues with remote
ZK clusters. Local colo connectivity is necessary for basic operations.

The degradation involves the following:
- No cross-colo requests from frontends and cross-colo replication lag at
  servers, leading to absence of read-after-write guarantees.
- Absence of discovering sealed state of partitions if something changes
  in remote colo (this can happen even when connectivity is lost after
  initialization today).

However, this is a small price to pay given that nodes can continue to start
up, intialize and be available for requests, and process most requests (all
PUT requests, all GET requests for older blobs and all GET requests for recent
blobs in the local colo.

Fixes linkedin/ambry#1069