Introduce request handlers for account metadata
Introduce a GET /accounts and a POST /accounts API. These allow users
of ambry to update account API through the frontend, as opposed to
writing there own thick client that instantiates the AccountService
class.
- GET /accounts can either return a collection of all registered account
  objects, or can fetch a single account by ID or name.
- POST /accounts takes a body that contains a list of accounts to add or
  update. Accounts cannot currently be removed using this API call.

This also includes some refactoring of AmbryBlobStorageService and the
related handlers to unify the common request preprocessing logic.