Fix size calculation in serializeString
For certain combinations of encoding and input string, String::length
returns the wrong size for the buffer that follows. One example is
encoding a String with multi-codepoint unicode characters (e.g. emoji)
into ASCII. In this case, the encoder will replace the multi-codepoint
characters with a single "?" character. This prevents such records
from being successfully deserialized.