Introducing AmbryStatsReport to support multiple types of stats
Previously we use QuotaHealthReport to encapsulate all stats for a
single node. As we are introducing more types of stats report, this will
make single znode complicated and exceed size limitation. In this PR, a
new AmbryStatsReport is introduced to replace QuotaHealthReport, thus,
we can have separate stats reports for a single node and aggregate them
separately.