Allow NettyServer to use its own ssl impl (#1117)
Previously, the Netty code was forced to use the same SSL impl as the
router. This adds a config to let NettyServer use a different SSLFactory
implementation. This can be used to leverage the performance benefits of
NettySslFactory within the HTTP server, without affecting the router,
which has some compatability issues with Netty's OpenSSL-based
SSLEngine.