Avoid returning 416 when range end past blob
The original implementation of ByteRange.toResolvedByteRange was
stricter than the RFC called for and required that the end of the range
be less than the blob size. This change allows ranges where the end
offset or suffix length is greater than the blob size to be served.

Per rfc7233:
   A client can limit the number of bytes requested without knowing the
   size of the selected representation.  If the last-byte-pos value is
   absent, or if the value is greater than or equal to the current
   length of the representation data, the byte range is interpreted as
   the remainder of the representation
and
   If the selected representation is shorter than the specified
   suffix-length, the entire representation is used.

This change also refactors the implementation of ByteRange to use
inheritance instead of switch statements.