Make Bootstrap tool use Helix Cluster Manager for validation (#1116)
Instantiate a Helix cluster manager within the bootstrap tool to listen
on the changes that it makes itself. This cluster manager's state will
be checked during validation. If there are other errors that can cause
errors such as Out of Memory in the Helix client, then those can be
immediately detected because the tool itself will be affected.