Fix intermittent index test failure
The test failure is due to Store I/O error was not always triggered when
invoking getBlobReadInfo(). The reason is, in getDeletedBlobReadOptions,
the hardDelete.getMessageInfo() is called only when PUT and DELETE are
in the same log segment. This PR explictly maintain one deleted key
whose associated PUT record is within same log segment to ensure test
succeeds.