Introduce container metrics
Introduce request metrics scoped to individual containers for a subset
of the operations. These will be created lazily when a request comes for
a container to limit the number of metrics created for rarely used
containers.
Refactor RestRequestMetricsGroup to make it usable for metrics without
encrypted variants.
Remove completely unused legacy metrics interfaces.