Fixing the intermittent Helix tool test failure (#1162)
The replica size is changed in testEverything() method. If new replica
from initial partition is added to new node disk and is not reset when
calling writeBootstrapOrUpgrade(), then the second time it is invoked,
validating cluster manager in Helix tool would see different size of
replica of same partition and fails the test.
This patch ensures that if new replica is from initial partition, the
replica is only allowed to be added to original nodes in HardwareLayout,
which would be reset by writeBootstrapOrUpgrade() and avoid inconsistent
replica size error in validating cluster manager.