Free PutOperation references earlier
This contains a couple of fixes to free references to PutOperations
earlier and to allow GC to take place:

Since NettyMessageProcessor keeps a reference to the last NettyRequest
object sent over an HTTP connection, there can be a link to a callback
that prevents a previous PutOperation's buffer from being GCed. The
easiest way to address this is to nullify the callback reference after
it is used.

correlationIdToPutOperation map would previously retain a reference to
a PutOperation even if the operation was fully complete. This means that
large buffers would be retained until all responses are received, even
if the server response is no longer relevant to the operation.