Set permissions for ambry data and operation files
For security concern, data files and operation files are required to be
set with specified permissions. For example, user data related files
should be set "rw-rw----" and operation files (usually generated by
Ambry itself and don't contain user data/metadata) are allowed to set
"rw-rw-r--". The permissions are configurable and could be changed in
StoreConfig.