Upload new added replicas info onto Helix PropertyStore
This PR supports uploading infos of new added replicas onto Helix
PropertyStore. This is part of move replica which specifies the location
of new replicas and their size, partition class etc. The Helix bootstrap
upgrade tool will extract diff between static clustermap and clustermap
in Helix. The diff is used to generate a map of new added replicas and
then upload it to Helix.