We are seeing some build failures in dev environment since "org.gradle.caching=true" has been set in gradle.properties. (#1255)
Some of the failures have been because gradle seems to cache a module if no existing files are modified in the module, and doesn't rebuild when new files are added.
We should set org.gradle.caching to false to fix this. Setting this to true doesn't buy us a lot of performance, because the compilation is fast anyways.