Add chunk ID to get exceptions, add max in-mem chunks config
Add the chunk ID to the exception messages for any exceptions that
occur when fetching a data chunk in GetBlobOperation.

Add config for the max number of data chunks to keep in memory for
a single PutOperations or GetBlobOperation. This also determines
the max number of chunk requests that may be in flight to a server
node in a single operation.