Close connections on router request timeouts
This change provides a way for the router to indicate to the
NetworkClient that a request should be dropped (currently implemented by
closing the connection that the request was sent over). Allowing the
NetworkClient to kill unhealthy connections and reinit them should
prevent the connection pool from being unnecessarily occupied by failed
requests that the router no longer cares about.

This change requires informing the NetworkClient about request
correlation IDs so that there is a key to drop a request by (as the
rest of the router does not know about connection IDs).