Add max replenishment per host per second config
This config will provide a way to limit the number of connections
created per second by connection replenishment to each host. With this
change, replenishConnections will only be called once per second (in
this case, a good enough approximation that avoids the need for a
per-host token bucket).

This is meant to prevent connect/handshake floods when servers are
restarted in clusters with many frontends.